Merged PR 21047: SS_TICKET_PICKUP missing param
A parameter that we used to test workflow behavior in container was not merged, by accident.
Re-adding here and self-approving as need this in a release sunday evening and there is no risk as this action is currently in use by no customers. (Monday will be the first)

Related work items: #10945