Merged PR 21188: NETS_CLOUD fail on some transactions in BC but not on terminal
The Response from NETS, sometimes give a null value on the OptionalData field, which the solution could not handle. So a safe wrapper have been added to this.
This is the only know Json property know to cause the issue so it is only this field we do it on.