Merged PR 21037: Feature Flags
New functionality - Feature Flags

The functionality allows the developers to configure and use feature flags that can be enabled/disabled remotely in the customer's environments. The feature flags are hosted in ConfigCat and can be edited there.

 - codeunit 6151499 "NPR Get Feature Flags JQ"
The codeunit is set up as a job queue and it runs every one minute. It downloads the configuration from ConfigCat to a table - "Feature Flags".

- Function IsEnabled in codeunit 6151495 "NPR Feature Flags Management" can be used to check if the feature flag has been enabled.

Related work items: #14245