Merged PR 21250: Async Order Posting after modifications
When buffer item lines are created in POS Entry SL they represent the latest version of document lines in that moment. When JQ tries to post that document it will compare if Amounts, Quantities, Unit Price,Variant Code in document lines are the same as in buffer lines in POS Entry SL. If they are not the same, JQ will raise an error and posting will fail.
Before these changes comparison was done against buffer lines of initial document pos entry (usually an order) and not against the latest buffer lines. So, if prepayments are made the code would check if order lines are the same as they were when order was created. This was wrong as user could change the order in the meantime.
Now , JQ will compare document lines with buffer lines from the current transaction.
POS Entry Sales line Relation table is added to ensure 1:1 relationship between buffer and document lines.
It will be deleted when order is fully posted.

Posting can also fail because of missing values in the document (eg. Variant Code). In that case (when posting status is Error) user should be able to change the document. If the document is changed there is no point in lines comparison.  JQ will post the document as it is at that moment ,without checks.

In codeunit 6014407 "NPR Sales Doc. Exp. Mgt." CalcInvoice is removed because "Qty. to Invoice" will be filled based on Sales and Receivables Setup. So it shoudn't be taken into account.

Related work items: #14274