Merged PR 21280: POS Actions not named for script injection
To ensure automatically minifying and injecting the contents of any .js file into an AL string those two files needs to be in the same folder and that their names start with POSAction.
So this PR is about renaming and moving a large number of objects.
The exceptions are public objects and objects with published events.

Related work items: #12938