Merged PR 21282: ISV1-23 Mixed Discount Performance Improvements
The mix discount calculation was very slow when you had a mix discount with thousands of lines. The main reason for that was that there were several functions that were looping through all of the discount lines for different reasons one of which was to check if the discount was applicable to the current POS Sale. The behavior of those functions was altered, and an attempt was made to limit their use.
Support was added for Feature flags. The feature flag that enables/disables the new calculation is newMixDiscountCalculation.