PR 21323: Retention policy job set to run multiple times each night
Due to:
- the recent change done by Microsoft to the maximum number of records that can be deleted during a single retention policy job run (from 250K to 10K in BC 22.5), and
- the fact that standard retention policy job can pretty frequently end with an `Arithmetic operation resulted in an overflow.` error (at the same time some records gets deleted though)

we needed to update the way retention policy job is going to be run for our customers.
From now retention policy is going to be run as many times as possible every night (during allotted time between 23:30 and 06:00). Running it only once per night (as it used to run) doesn't make any sense, as deleting of total of 10K records a day is absolutely not enough.

This will need to be changed as soon as Microsoft comes up with a more stable and reliable solution.

Linear issue: [ISV2-30](https://linear.app/navipartner/issue/ISV2-30)
Related work items: #14629