Merged PR 21344: Linear ISV1-32 Separation of total discount calculation from other ad hoc discounts
In the past, the buttons remove line discount and remove total discount were able to remove any type of discount on the POS. After the introduction of the total discount, this behaviour was altered because any change in the POS sale triggered a total transaction recalculation. This was causing the calculation of mix, quantity and any other type of ad hoc discount to be triggered and the discounts to be calculated again after the line or total discount was removed from the POS Sale. The end result was that both buttons were able to remove only manual discounts.

To correct this behavior the total discount calculation was separated from the calculation of the ad hoc discounts. Now you can edit the transaction as much as you want and the total discount is going to be applied on the edited POS sale without triggering the ad hoc discount calculation.