Merged PR 21424: Activity Coupon
A new coupon application type has been developed. The idea of the coupon is that the company will issue just one coupon with a specific reference code that can be used multiple times and will be printed on newspapers, flyers, etc. When a customer acquires the coupon, he can use it to get a discount on 1 unique unit for a specific item. This means that only 1 quantity can be discounted from a unique item.  If the user has two coupons, then the coupon will be able to discount a second unique unit from an item - you can discount only 1 quantity from an item that has not already been discounted. If we have items with the same item no but different serial numbers, then these items are going to be treated as completely different items.