Merged PR 21494: Disable abort on payment transaction dialog
After consulting with Pepper I decided to disable the abort button on the transaction dialog.

From: Pepper <pepper@treibauf.ch>
Sent: den 24 februari 2022 16:44
To: Tim Sannes <tsa@navipartner.dk>
Subject: Re: [Ticket#0120909] Hi Suppport,

Dear Tim Sannes

This utility code only can work if the terminal supports it. This a functionality that won’t be supported anymore in Pepper (22.1) and will stay in Legacy Pepper.
The reason is that most modern terminal don’t support this anymore is due to PCI and Security reasons.

Otherwise the only safe way to stop a started transaction is by pressing the cancel button on the terminal.
If you have any further questions, don’t hesitate to ask.
Best regards

Lauren Arekhi
Pepper Support
T +41 44 355 50 83
E pepper@treibauf.ch
treibauf AG, Gubelstrasse 59, 8050 Zürich
www.treibauf.ch

24.02.2022 12:00 - Tim Sannes schrieb:

I have a question regarding eftUtility and opcode 0 for Aborting the started transaction. This is in regards to legacy (19.1 f.ex.).

If I call eftUtility with the abort opcode on an ongoing EftTrx transaction, the transactions fails (with -30). Is this how it is intended to be used?
(Also the eftTrx result data is garbled – like it is read from un-initialized memory). Do I need to specify extra information in the XML parameter?

I imagined that the POS starts a transaction and is waiting for the customer to swipe the card. But at this moment the customer change his mind and would rather pay with cash.
You could click the cancel button on the terminal – but I thought abort would be able to nicely stop the ongoing transaction – nothing has really happened.

Please advise. How do I ask the terminal to nicely attempt to abort the ongoing transaction?