Merged PR 20865: Insert Item Performance Improvements
We have done some performance improvements when inserting an item in a POS Sale. Before the change following workflows were called:

- TEXT_ENTER - this workflow finds if the code that we're inserting in the POS Sale is an item or something else.
- ITEM - After TEXT_ENTER has found out that we're actually inserting an item then the ITEM workflow was called.

After the changes, the flow is the following:
If TEXT_ENTER finds that we're inserting an item, then it checks if input is needed from the user. If no input is needed, then TEXT_ENTER directly inserts the item in the POS Sale without calling the ITEM Workflow. If input is needed from the user then TEXT_ENTER calls the ITEM workflow.

The new feature supports Feature Flags. It can be enabled/disabled by the feature flag - textEnterFastItem.

Related work items: #14106