Merged PR 21513: OAuth Token refresh
Afaics, the main reason for OAuth Setup table is for the CnC module - sending an order from one company to another with POST method.

This table looks like a nice patch in cases in which POST method needs to be sent from BC and in which an app doesn't really need to be registered in Azure AD in BC.  It can be used as a generic solution for PTEs -Oauth setup code can be added to a master table and with 2 subsribers everything will be ready.

Client secret and ID are something that should be stored (it will be stored only once) but access token and its lifetime can be cached.
Page action are hidden as every user would have their own instance, i can't see the point in caching it from the page.

Linear issue for more details:
https://linear.app/navipartner/issue/ISV2-41/oauth-token-refresh-inside-tryfunction

Related work items: #14688