Merged PR 21607: Refresh the item parameters of the POS Named Actions Profiles after upgrade
We have added a new parameter on the Item action. Unfortunately not all named actions profiles got updated with the new parameter. The only ones that were updated were the ones with the blank code.

The current pr creates an upgrade procedure which loops through the POS Named Action Parameters and checks if the parameter of the item action need to be refreshed. If needed they are refreshed.