Merged PR 21630: Upgrade variety in "Magento Picture Link" bugfix
While I was looking into 597802 to see what the issue is, I have found bug for upgrade codeunit I have created based on case 533149 and PR: https://navipartner.visualstudio.com/NpCore/_git/NpCore/pullrequest/20294 for variety cleanup in "Magento Picture Link".

It was OK initially, I tested it, but than we have had some changes in the logic, so I had to change and test it again. And in order to run same upgrade, I had to change upgrade tag to something else...
After second test, I have returned it back in tags, but forgot to return it in codeunit itself. :( I appologize.

This PR is now fix for that stupid error.

P.S. I think that we should create one more PR after this one, and change **Pending** to **Obsoleted** on field "Variant Value Code", and let it be included into new release at the end of the month. Based on that, we will have an error in the future, so consultants will know that they should change XML templates.

Also, while I was checking this, I have noticed that field "Variety Value" to which we moved the data, has table relation to a table "NPR Variety Value" but when we are inserting data, we are not validating it and that is causing dead relations. It's not a big deal, it can go like that, but it also could be fixed...