Merged PR 21727: Retention policy codeunit and report refactoring
Retention policy codeunit and report refactoring
Report for updating retention policy was separately developed and it was not in sync with codeunit used for updating the retention policy on upgrade. Now whole process of updating of retention policy is in one codeunit and with new parameter is added to determine if this is updated on upgrade or not.
[ISV2-60](https://linear.app/navipartner/issue/ISV2-60/retention-policy-update-report-is-not-in-sync-with-retention-policy)