Merged PR 21733: Gift voucher scanning problem
There was a problem when scanning the barcode on the Gift Voucher Report - the barcode reader was not able to read. The problem was a combination of the barcode colour and the barcode type. Changing the colour of the barcode to black and the barcode type to code 39 resolved the issue.