PR 21788: Adjusted retention policy job starting time
It turned out that the standard Job Queue rescheduling functionality does not understand overtime jobs, when the starting time is defined as a late hour of one day and the ending time is defined as an early hour of the next day, and we need the job to run as many times as possible during the allotted period between the starting and the ending time. Instead system just use the time period between the starting time and the end of the day (24:00), and then reschedules the job for the next day's starting time.
For example, retention policy job has had the `Starting Time` defined as 23:30, and the `Ending Time` defined as 06:00, with `No. of Minutes Between Runs` = 2. As a result the job would just run between 23:30 and 00:00, and then would be rescheduled for the next day's 23:30.

With this update we are changing the starting time for the job to 00:00. This should avoid the drawback of the standard job queue rescheduling routine.

Linear issue: [ISV2-135](https://linear.app/navipartner/issue/ISV2-135)