Merged PR 21796: Select from voucher list does not work in action SCAN_VOUCHER_2
We have a setup that allows us to open the voucher list instead of scanning the voucher no/barcode.

The function is supposed to work, when we do enter on the blank text field of the pop-up that we use for scanning voucher no. Hence, instead of scanning, we call for the list of vouchers.