Merged PR 21941: Item Worksheet Registration problem when using language different than english
When trying to register an item from the item worksheet if the language of Business Central was different than English there was a problem when checking the "item type" field value - the type field in the worksheet is not translated, but the item type in the item card is translated. So when Business Central was comparing the values it was comparing the captions of the values which were different because of the translation and it was deciding that there was a change in the value of the field.

The current pr tries to fix this behavior.