Merged PR 21905: Stripe payment gateway
When online payment is made and PaymentIntent object created, Stripe payment ID must be used for capture, refund or cancellation. Stripe only facilitates the payment, and does not have access to information relating to the fulfilment of orders.
OrderId can be sent as optional parameter, but it can't be used for payment handling.
Therefore we need to have the Stripe Payment ID stored in order to use it later for capture,cancellation or refund. PaymentIntent ID is caps sensitive.
https://stripe.com/docs/api/payment_intents/capture
https://stripe.com/docs/api/payment_intents/cancel
https://stripe.com/docs/api/refunds/create

Related work items: #14821