Merged PR 21786: Retention policy for NPR Nc Task
Standard way of deleting records when filters are defined have bad performance because of the overlapping data sets on creation of pool of records to be deleted. New implementation is created for the table NPR Nc Task, prevention for modify and delete of NPR filters is set. Deletion will be performed in linear way, filters from Retention policy line are set on the record and after that deletion is performed, this repeats for every NPR defined line.
[ISV2-99](https://linear.app/navipartner/issue/ISV2-99/adjust-retention-times)