Merged PR 22032: End Sale Workflow
We have created a new workflow - END_ SALE. This workflow will be used to end the POS Sales. The idea is to have one unified point in the POS workflow to end the sale in order for integration to be able hook into that workflow and work based on the information that we return to them. The END_SALE will support all v3 actions except the SS actions.

The workflow has been used to end the sale in the following workflows:
- PAYMENT_2
- ISSUE_RETURN_VCHR_2
- SCAN_VOUCHER_2

A flag has been created in configcat - endSaleWorkflowEnabled - that enables the use of the new workflow.