Merged PR 22138: Get Feature Flags job log in scheduling issue
We have developed a functionality that resets the feature flag job once a day if the job is not running. Unfortunately, the functionality was creating the job queue if it didn't exist in the company. This is an issue, because we want the job to run in only one company per environment.  The pr addresses this issue.