Merged PR 22159: Callback for POS Stores and Units update change
Changes in callback to case system for updating POS Stores and Units in BC Service Tier Tenant and BC SaaS Tenant.

Current update is working like Mark asked more than a year ago. Then there was one change a month ago, when he asked to filter counting by status (open/closed, active/inactive).

In the past there was many "check quantities" cases and every time, it wasn't problem in the code.
Based on the requests and experience we are now introducing new logic.

Quantities will be updated when:
New store or unit is inserted (but on the next login, regardless when the last sync was made)
New store or unit is deleted (same as above)
Status of store or unit is modified (same as above)
And on after user login, but only if the last sync was 24 hours ago (or more). So, in theory, we could have one sync per company, per day.

I took my time for this change. I have tested it on case container for NAV Service Tier and for BC SaaS
I have tested all possible scenarios.
Insert, Delete, Modify (with and without changing the status), and not changing anything, but just resync after 24 hours.
I have also tried to delete the data in case system and confirmed that it will be recreated during next sync. (that is scenario if someone accidentally delete data from case system).