Merged PR 22288: New Gift Voucher Reports that uses CODE128 barcodes
At the moment the gift voucher reports use code 39 format to generate the barcode on them. Some customers are using this format but code 128 is the default format and is used by most people so we need to support it. After some back-and-forth changeing of the reports, we have decided to have two sets of reports - one that uses Code39 and one that uses Code128. Because of this, we have created copies of the gift voucher reports that use code 128 for generating the codes.