Merged PR 22194: New custom voucher reference no experience
We have changed the flow when a voucher with a customer reference no is issued. When the voucher is issued a popup is going to appear where the salesperson will be able to scan the custom reference no. Each voucher is going to be on a separate line.

A new pos action was created - SET_VOUCHER_REF_NO that can be used to change the reference no of the voucher.