Merged PR 22325: Text formatting mismatch fix
Text formatting mismatch fix
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/22325/attachments/image.png)
Error pops up on  the Retail Manager Rolecenter on the  customer production environment. This was due the regional formatting of the numbers when inserting into Chart Data table and after that Evaluating results for populating the Chart Buffer. Now formatting is changed to be in XML format, reading of data is performed with the check to prevent future errors.