Merged PR 22345: App area enabling on company creation
App area enabling on company creation
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/22345/attachments/image.png)
Problem occurs only on the SaaS environments when creating the company using action available on the page. After company is created and user logins the wizard  buttons are not showing. This was because NPRetail Application Area was not enabled on company creation. Added the call to the RefreshExperienceTierCurrentCompany procedure which will enable the application areas if it's Enabled in NPR Feature Management on Company Initialization.
[ISV2-167](https://linear.app/navipartner/issue/ISV2-167/nprretail-application-area-is-not-activated-for-newly-created)