Merged PR 22481: Integration event for changing field number filter on page
Integration event for changing field number filter on page
[ISV2-203](https://linear.app/navipartner/issue/ISV2-203/upgrade-nav-2009-to-saas-updating-prices-and-creating-new-items-600315)
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/22481/attachments/image.png)
Field "Base Unit of Measure" (22) from table 6060042 "NPR Item Worksheet Line" is not available on the page Item Worksheet Field Setup for lookup, this is due restrictions in NpCore, filter is set to exclude this field in Lookup Page. Integration event is added to make this available.