Merged PR 22560: #com-215 Error when one of the card payments is not successful #com-216 Investigate json format file error
POS Entry Payment Lines where Amount is 0 due to not successful payment with the card should be skipped, since fiscal printer does not accept 0 amount in Json for payments.

Absolut value for POS Entry Payment Line Amount should not be sent (just as it is) in order to handle change return to customer.

https://linear.app/navipartner/issue/COM-215/error-when-one-of-the-card-payments-is-not-successful

https://linear.app/navipartner/issue/COM-216/investigate-json-format-file-error