Merged PR 22615: Generate Posting No. for Async Invoices
Generate Posting No. for Reporting purposes

After Sale Document is posted Bravo HU will send it to the authorities. Among other details, the report will show the No. of posted document.
The same report should be printed to the customer when the sales document is posted on the POS at the end of the POS sale. With async posting we don't have Posting No. value in the real time.

"Generate Posting No." is added in order to provide Posting No. in the real time to the docs with async posting. System will reserve and assign Posting No. that will be taken into account later with JQ post.
Posting No. will be generated only for Final invoces and credit memos, that are fully posted.

Related work items: #15592