Merged PR 22618: fix for multiple tickets on mPOS printing
fix for multiple tickets on mPOS printing.
When printing for mPOS, it used the controladdin page for MPOS Connector in a Modal page.
Becuase the Ticket Record is modified with Printed Date, then any subsequent prints would fail due to the opening of a Modal Page ends the write transaction.