Merged PR 22423: Changed how admission schedules are generated
In light of some corner cases where manual editing of schedule entry data is overwritten by schedular defaults when schedules is checked and re-generated to be up-to-date.

The individual change requests are handled individually by type
* New (insert)
* Delete (update existing to status cancel)
* Update Replace (New + Delete)
* Update Small - when user has changed "editable" data. Updates with Schedule times.

Update replace occurs when schedule is manipulated such as adding or removing schedule lines
Update small occurs when user has edited the entry (capacity f.ex.) and the scheduler time changes.

Added a bunch of test scenarios.