Merged PR 22700: Membership Rewards
This is new functionality - a complement to loyalty.

You setup up a goal and certain activities will increase your score for that goal. There may be multiple activities for the same goal and each activity has its own weight and conditions. Each time an activity is detected, the membership score will increase with the activity weight. All activities are evaluated and multiple activities can apply at the same time. Activities has start and stop dates

A goal has a threshold level. When the goal reaches the threshold, the goal has been achieved and a reward can be issued (coupon).

One goal has one threshold. But goals can depend on other achievements before you start earning on them. Goals have start and stop dates.

Current activities are:
* Member Arrival. When you swipe your member card and there is a admission created. Can be filtered on AdmissionCode.
* Member Achievement. When a goal is reached and achievement is accomplished. Can be filtered on GoalCode
* Manual. Admin can manually add or decrease the score for a goal.
* NoOp. A dud.

In addition, activities has some general conditions:
* Frequency: Activity can only occur once per time interval
  - CY, CM, CW, CD - Current [Year, Month, Week, Day] or blank
* Weekday: Activity must occur on a specific weekday
  - WD1..WD7 (1=Sunday, 2=Monday, ...) or blank for any day
* Month
  - M1..M12 (1=January, ... ) or blank for any month.

A condition for an activity for some goal could read as: "You will be awarded 10 points for visiting admission X on a Monday in May, once."

Activities are additive, so if there was some other activity without constraints for the same goal with let's say 1 point, membership would be awarded 11 points on that one Monday in May, but only 1 point every other day.

* "5 visits in a Season"
* "5 visits in a Seasons on _different week days_":
(1 goal with threshold 5 and 7 activities, each activity would only be valid on a specific weekday)

**Note: it is not the points themself that are have value, it is reaching the threshold that triggers the reward.**