Merged PR 22718: Assembly order not created with SALE_DOC_EXP
When a POS Sale that contains a BOM Item that is set to assemble to order was exported to a sales order, the system was not creating the Assembly order for the BOM. This was caused by the fact that we were validating the quantity before the sales line was created and because of that the system was not seeing the sales line and was not creating the assembly order. We have moved the quantity to assembly validation to be after the sales line creation.