PR 22754: Bin Transfer: direct bin transfer between POS units
System will now auto-create bin transfer IN journal line for POS2 after the corresponding "Transfer OUT" transaction has been posted on POS1
The process is as follows:
- POS1 creates a bin transfer OUT transaction, and mentions POS2 as the final recipient.
- When the transfer OUT transaction is finished, system automatically prepares and releases a bin transfer journal line for POS2.
- On the last step POS2 simply needs to start a transfer IN transaction, get the prepared bin transfer journal line and confirm it (no need to manually fill in any fields, and that will prevent users from making errors).

An In-Transit bin (of type "Safe") is used as the bin through which the amount is posted for both IN and OUT transfers. There's a new field on the "POS End of Day/Bin Tr. Profile Card" page where this intermediary bin code must be specified. The field name is "In-Transit Bin Code"
![image.png](https://dev.azure.com/navipartner/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/22754/attachments/image.png)
If the field is empty, users won't be able to select other POS units' bins (bins of type "cash-drawer") on transfer OUT transactions.

Linear issue: [ISV2-225](https://linear.app/navipartner/issue/ISV2-225)