Merged PR 21785: Fiskaly Digital Receipt
[ISV1-28](https://linear.app/navipartner/issue/ISV1-28/digital-receipt) Fiskaly Digital Receipt

This new feature will allow POS to issue a digital receipt after the end of sale.
Added new table/page 'POS Receipt Profile' which holds setup for this to work and which will be connected with POS via POS Unit(it has profile field on POS Unit card). On POS Receipt profile we can select whether to enable or disable issuing of digital receipt as well as timeout enabling which will be explained later on.
All receipts are being stored in POS Sale Digital Receipt Entry table.
At the end of the sale digital receipt is being issued as well as QR Code being shown on POS in pop-up window. Pop-up can be closed manually, but also we can set timeout(in seconds) on POS Receipt Profile which will auto-close this pop-up if specified.
Another feature on setup page is showing of discount lines based on the option selected: per line, summary and no information(which are self-explanatory I assume).
We also have 2 factboxes on the POS Sale Digital Receipt Entries page which are visualizing both PDF of generated digital receipt and its QR Code