Merged PR 22839: Total Discount Optimizations
The total discount was taking time on all payments. This was happening because we were always calling workflow CALCULATE_DISCOUNTS which was check if there is an active total discount and it was applying to the sale. Since calling workflows always adds an overhead to the executing time we're now checking if there if there is an active total discount for today before calling the workflow.

We have added new events for the total discount. It needs to run separately from the other discounts so there was no point in using the same events as the other discount types. They were adding overhead time to the calculation of the total discount because each time when the total discount was called we were checking if there were other active type of discounts. To resolve this we've created new events that should be used only by discounts that have a similar behavior as the total discount.

Additional optimizations were done to some of the procedures used by the total discount.