PR 22874: Retail vouchers: prevent duplicate reference numbers
This is in addition to this PR: !16624
When assigning a reference number for a new retail voucher, the system didn't check if the same reference number had already been used for an archived voucher. It only checked for active vouchers.
This is a problem when we tried to integrate retail vouchers with Shopify for Vibholm.

Related work items: #9885