Merged PR 22862: Fix mix discount applies negative discount amount when total Amount per min. qty used
The mix discount was applying a negative discount amount when Total Amount per Min. Qty option was used and the total amount of the transaction was lower than the total amount specified in the mix discount. It was doing that to apply the amount specified in the mix discount. We have made a change to skip the discount amount application in this case.