Merged PR 23058: Changes in callback to case system. Sync POS Stores and Units every 24 hours,...
Changes in callback to case system. Sync POS Stores and Units:

- every 24 hours, and not only when there was a change, how it was till now.
- when company information is changed from PROD to DEMO or TESTING, send 0 quantities to case system
- when company information is changed from DEMO or TESTING to PROD, count and send actual quantities to case system
- when company is deleted, send 0 quantities to case system

Also, there is additional field being sent to case system to support storing same name companies in multiple environments. For example, LEGO has Bulgaria, Croatia, Slovenia and Serbia environments and every single of those has company named "My Company". And all of them are on the same Azure AD Tenant, so we could not record them without Environment Name.

Because of this, new azure function (UpdateSaasPosStoresAndUnitsV2) is created, while UpdateSaasPosStoresAndUnits is still there, to support tenants which are not upgraded to version after this PR.

Code in case system is also changed to accept new parameters. (added field Environment Name)
And primary key of table "BC SaaS Company" is changed also, so now it contain Environment Name field. And all references are updated to do GET and if failed, SETRANGE with Azure AD Tenant ID and Company Name, so it will support period until all clients are upgraded to new NP Retail with this PR in it.

I have tested everything many times and it's safe to go.