PR 23147: Table "Audit Roll" cleanup upgrade code disabled
There is an issue with the "Audit Roll" table cleanup code in codeunit 6151581 "NPR Obsolete Tables Cleanup". It seems to fail if there are a lot of records in the table. As a result, the whole upgrade process fails as well.

This PR disables the upgrade code in BC SaaS.
Either the table needs to be cleared before migrating to BC SaaS, or we need to find another way of clearing up the table.

Linear issue: [ISV2-271](https://linear.app/navipartner/issue/ISV2-271)