Merged PR 23002: [RS] Fiscal - Changed Sales Header Retail Conditions
**[RS] Fiscal - Changed Sales Header Retail Conditions**

Added the check for retail location on Sales Lines as a procedure that, for standard Sales Order, gets called OnBeforeReleaseManualSalesDoc and sets the global var, so in OnAfterReleaseManualSalesDoc and OnAfterPostSalesDoc it doesn't have to check it all again.

Added the check for Prepayment posting as well, since there was an issue with posting the prepayment invoice to non-retail location - Error for RS Aux Headers popped up, when it shouldn't have.