Merged PR 23192: Coupon VAT Calculation Fix
There was a problem when the "Price Includes VAT" checkbox on the sales line was set to false. The amount on the issue coupon line in the coupon entries included VAT and the amount on the discount application line didn't include VAT. We have made changes so that the amount on the discount application line includes vat as well.

https://linear.app/navipartner/issue/ISV1-165/coupons-discrepancy-upon-redeeming-loyalty-points