Merged PR 23249: Skip simple insert if there is postworkflow subscription
The simple insert should be be skipped if there is an item insert postworkflow subscription. We were deciding if there is a postworkflow subscription, based on the postworkflows that were returned by the OnAddPostWorkflowsToRun event publisher. This proved not to be that good and now we're checking if there are event subscriptions for that event in the event subscription table instead.