Merged PR 23372: Added Standard Exchange Rate func on POS and fixed Balancing Lines calc
Added Standard Exchange Rate func on POS and fixed Balancing Lines calc
[ISV1-256](https://linear.app/navipartner/issue/ISV1-256/end-of-day-balance-with-currencies-uses-different-exchange-rate-than) [ISV1-243](https://linear.app/navipartner/issue/ISV1-243/pos-payment-method-use-standard-exchange-rate-from-bc-has-no)

We were calculating amounts(FCY and LCY) only with Fixed Rate on POS Payment Method. With these changes button 'Use Stand. Exc. Rate for Bal.' has purpose. If it is on -> the amounts are calculated with Currency Standard Exchange Rate. If it is off it is using our Fixed Rate field as before.
Calculations of Balancing Lines are also fixed according to this