Merged PR 23542: Shipping provider integration Shipmondo Packeta improvements
The shipping provider Packeta supported by Shipmondo requires that we send the value of the shipment. For that we have added a new field on the "Package Shipping Agents" - "Declared Value Required" if set to true we will send the amount of the shipping goods.

We have changed the way the address1 and address2 are sent, because Packeta has a limitation on the number of symbols on the address 1 field and at the same time requires that address 1 field contains the number of the address.

We were using the Package Shipping Agents". Code to create the products that Shipmondo supports, but the field length was too small. We have created a new field Shipping Provider Code which is going to be used by the Shipmondo integration. It will be initialized with the value from the Code fields, but the user can change it if necessary.

We have added fields for "Default Length", "Default Height", "Default Width".

https://linear.app/navipartner/issue/ISV1-274/shipping-provider-issue-with-code-filed-length-code-1-code[10]-pk

https://linear.app/navipartner/issue/ISV1-279/shipping-provider-error-shipping-provider-documents-send-document-lego

https://linear.app/navipartner/issue/ISV1-280/shipping-provider-shipment-method-code-agent-agent-service-validation