NP Shopify integration (#7162)
After considering all the pros and cons, we have decided that we should
continue to use the NaviPartner Shopify integration functionality,
rather than switching to the standard MS Shopify connector app.
This PR is about moving the functionality we had previously in the
NaviPartner Shopify integration app to Np Retail. I have merged what we
had in Vibholm with our “standard” we had for other customers and
included everything in Np Core.