KDS API methods moved out of dragonglass into a separate API codeunit (#7179)
Originally we used the Dragonglass custom method interface to call KDS
API methods.
This approach was a bit slower as it needed to initialise a bunch of
unrelated POS objects.
I have now created a separate codeunit for the KDS API methods, which
will be published as a separate web service.

I have also moved out all of the KDS related objects to a separate
folder