Skip restaurant flow if there are no items to produce (#7197)
The system should skip the restaurant flow, if the POS sale does not
contain items to be produced for any of the restaurant kitchen stations.

Also fixed restaurant SMS notification sending to a group of recipients
specified in the SMS template. Now the group can also contain only one
recipient.