Handling of waiter pad lines with negative producible quantity (#7243)
In KDS environment:
If there was an item line on a waiter pad with a negative producible
quantity (typically a modification of another (main) item line), the
system would not automatically close the waiter pad. This was because
the system was expecting the line to be served (and it would never be
marked as served because kitchen requests aren't created for waiter pad
lines with negative quantities).

This PR fixes the problem. Waiter pad lines with negative quantities
should no longer prevent the system from automatically closing the
waiter pads.