Change item list coupon application to apply coupons to item that are not discounted (#7255)
We have identified an issue with the item list coupon application where
coupons consistently apply to the first item in the list, regardless of
whether the amount, including VAT, is zero. Moving forward, we are
implementing a change in the system's behavior: coupons will now apply
to the line with the highest amount, including VAT. This adjustment
takes into account all other setups for priority, highest, and lowest
price application.