Fix logic in MMNPRMembership.Codeunit.al for handling ExpiryDate (#7274)
The commit fixes the logic in the MMNPRMembership.Codeunit.al file where
the ExpiryDate is checked. The change ensures that if the ExpiryDate is
either greater than or equal to Today's date or if it is 0D (empty), and
the Blocked flag is false, then the TmpMemberInfoCapture record is
inserted.